[HIBIKE] Util.h Refactor
Creates new files util.h and util.cpp common to all devices.
Currently contains simple functions to manipulate the data buffer.

Removed redundant header file definitions of the hibike spec.
