+++ hibike/lib/hibike/hibike_device.h
@@ -3,6 +3,7 @@
 
 #include "hibike_message.h"
 #include "devices.h"
+#include "util.h"
 #include <stdint.h>
 // For size_t
 #include <string.h>
@@ -35,7 +36,6 @@ void hibike_setup();
 void hibike_loop();
 void toggleLED();
 
-//extern uint8_t device_data_update(int param, uint8_t* data_update_buf, size_t buf_len);
 extern uint32_t device_write(uint8_t param, uint8_t* data,size_t len);
 extern uint8_t device_data_update(int param, uint8_t* data_update_buf, size_t buf_len);
